// types/Messages.ts
export enum Message {
  UPDATE_PLAYER = 'update_player',
  UPDATE_PLAYER_NAME = 'update_player_name',
  READY_TO_CONNECT = 'ready_to_connect',
  DISCONNECT_STREAM = 'disconnect_stream',
  CONNECT_TO_COMPUTER = 'connect_to_computer',
  DISCONNECT_FROM_COMPUTER = 'disconnect_from_computer',
  STOP_SCREEN_SHARE = 'stop_screen_share',
  CONNECT_TO_WHITEBOARD = 'connect_to_whiteboard',
  DISCONNECT_FROM_WHITEBOARD = 'disconnect_from_whiteboard',
  VIDEO_CONNECTED = 'video_connected',
  VIDEO_DISCONNECTED = 'video_disconnected',
  CHAT = 'chat',
  ADD_CHAT_MESSAGE = 'add_chat_message',
  SEND_ROOM_DATA = 'send_room_data',
}